% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{acupuncture}
\alias{acupuncture}
\title{Meta-analysis: 29 randomized controlled studies (RCT) assessing the efficacy of acupuncture
treatments as complementary treatment in depression patients}
\format{
A dataframe with 29 rows and 11 columns. Each row represents study results, the columns are:
   \describe{
    \item{author_year}{Author and year.}
    \item{hedges_g}{changes in severity between intervention and control groups calculated using Hedges´g statistic}
    \item{std_err}{Standard Error of Hedges´g}
    \item{intervention}{treatment administered}
    \item{comparison}{control group treatment}
    \item{country}{origin country of the study}
    \item{sample_size}{total amount of patients per study}
    \item{number_treatments}{number of treatments received per study}
    \item{variation_acupuncture_points}{fixed: same acupuncture points used at each session; semi-fixed: some points pre-defined, some selected on the basis of the diagnosis/symptoms (location and amount); individualised: location and amount of points selected on basis of the diagnosis/symptoms}
    \item{number_acupuncture_points}{amount of acupuncture points for fixed-points-studies}
    \item{NICMAN}{NICMAN scale Points to evaluate the Quality of the administered acupuncture}
    \item{random_sequence_generation}{Risk of selection bias (Random sequence generation) low risk of bias: high, high risk: low, unclear: unclear}
    \item{allocation_concealment}{Risk of selection bias (allocation concealment) low risk of bias: high, high risk: low, unclear: unclear}
    \item{blinding_participants_personnel}{Risk of performance bias (blinding of participants and personnel) low risk of bias: high, high risk: low, unclear: unclear}
    \item{blinding_outcome_assessment}{Risk of detection bias (blinding oft outcome assessment) low risk of bias: high, high risk: low, unclear: unclear}
    \item{incomplete_outcome_data}{Risk of attrition bias (incomplete outcome data) low risk of bias: high, high risk: low, unclear: unclear}
    \item{selective_reporting}{Risk of reporting bias (selective reporting) low risk of bias: high, high risk: low, unclear: unclear}
    \item{other_bias}{Risk of other biases; low risk of bias: high, high risk: low, unclear: unclear}
    }
}
\source{
Armour M, Smith CA, Wang LQ, Naidoo D, Yang GY, MacPherson H, Lee MS, Hay P. Acupuncture for Depression: A Systematic Review and Meta-Analysis. J Clin Med. 2019 Jul 31;8(8):1140. doi: 10.3390/jcm8081140. PMID: 31370200; PMCID: PMC6722678.
}
\description{
Meta-analysis of 29 studies on the effect of different methods of acupuncture
Therapy for depression compared to usual care control groups by pooling data from RCTs.
}
\keyword{datasets}
