% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{colon_cancer}
\alias{colon_cancer}
\title{Meta-analysis: Real World Evidence in metastatic colorectal cancer, comparing antiangiogenic treatments with chemotherapy}
\format{
A dataframe with 13 rows and 6 columns. Each row represents study results, the columns are:
   \describe{
    \item{study}{Author and year.}
    \item{study_type}{randomized clinical trial or comparative/single-arm real-world-evidence}
    \item{pfs}{logarithm of hazard ratios of progression-free survival}
    \item{se_pfs}{standard error of pfs}
    \item{os}{logarithm of hazard ratios of overall survival}
    \item{se_os}{standard error of os}
    }
}
\source{
Wheaton L, Papanikos A, Thomas A, Bujkiewicz S. Using Bayesian Evidence Synthesis Methods to Incorporate Real-World Evidence in Surrogate Endpoint Evaluation. Medical Decision Making. 2023;43(5):539-552. doi:10.1177/0272989X231162852
}
\description{
Meta-analysis of 7 RCTs, 4 cRWE studies, and 2 matched sRWE studies evaluating
progression-free survival (PFS) as a surrogate endpoint to overall survival (OS)
in metastatic colorectal cancer (mCRC), comparing antiangiogenic treatments with chemotherapy.
}
\keyword{datasets}
