% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{covid19}
\alias{covid19}
\title{Meta-analysis: Observational studies assessing the impact of
risk factors on the severity and mortality of COVID-19 cases}
\format{
A dataframe with 89 rows and 12 columns. Each row represents study results, the columns are:
   \describe{
    \item{author}{Principal author and year of publication.}
    \item{endpoint}{Endoint: severity or mortality.}
    \item{risk.factor}{Possible risk factors: diabetes, hypertension, cardiovascular, ACE_ARB.}
    \item{event.e}{Number of events in the group with risk factor.}
    \item{n.e}{Number of patients in the group with risk factor.}
    \item{event.c}{Number of events in the group without risk factor.}
    \item{n.c}{Number of patients in the group with risk factor.}
    \item{design}{Study design: Case Series, Cross Sectional and Retrospective Cohort.}
    \item{TE}{Log Odds Ratio}
    \item{seTE}{Standard Error of the Log Odds Ratio}
    \item{logitPc}{Logit transformation of the proportion of events in the control group.}
    \item{N}{Total number of patients.}
    }
}
\source{
de Almeida-Pititto, B., Dualib, P.M., Zajdenverg, L. et al. Severity and mortality of COVID 19 in patients with diabetes, hypertension and cardiovascular disease: a meta-analysis. Diabetol Metab Syndr 12, 75 (2020). https://doi.org/10.1186/s13098-020-00586-4
}
\description{
Meta-analysis of 40 Observational Studies from PubMed, Cochrane Library and SciELO databases
that assessed the impact of diabetes, hypertension, cardiovascular disease,
and the use of ACEI/ARB on severity and mortality of COVID-19 cases.
}
\keyword{datasets}
