% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.b3lmeta.R
\name{diagnostic.b3lmeta}
\alias{diagnostic.b3lmeta}
\title{Diagnostic function for b3lmeta object in jarbes}
\usage{
\method{diagnostic}{b3lmeta}(
  object,
  post.p.value.cut = 0.05,
  study.names = NULL,
  size.forest = 0.4,
  lwd.forest = 0.2,
  shape.forest = 23,
  ...
)
}
\arguments{
\item{object}{The object generated by the function b3lmeta.}

\item{post.p.value.cut}{Posterior p-value cut point to assess outliers.}

\item{study.names}{Character vector containing names of the studies used.}

\item{size.forest}{Size of the center symbol mark in the forest-plot lines}

\item{lwd.forest}{Thickness of the lines in the forest-plot}

\item{shape.forest}{Type of symbol for the center mark in the forest-plot lines}

\item{...}{\dots}
}
\description{
This function performers an approximated Bayesian cross-validation for a b3lmeta object
}
