% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.bmeta.R
\name{diagnostic.bmeta}
\alias{diagnostic.bmeta}
\title{Diagnostic function for bmeta object in jarbes}
\usage{
\method{diagnostic}{bmeta}(
  object,
  post.p.value.cut = 0.05,
  median.w = 1.5,
  study.names = NULL,
  size.forest = 0.4,
  lwd.forest = 0.2,
  shape.forest = 23,
  ...
)
}
\arguments{
\item{object}{The object generated by the function bmeta.}

\item{post.p.value.cut}{Posterior p-value cut point to assess outliers.}

\item{median.w}{Change color if median of a weight > median.w. The default value is 1.5.}

\item{study.names}{Character vector containing names of the studies used.}

\item{size.forest}{Size of the center symbol mark in the forest-plot lines}

\item{lwd.forest}{Thickness of the lines in the forest-plot}

\item{shape.forest}{Type of symbol for the center mark in the forest-plot lines}

\item{...}{\dots}
}
\description{
This function performers an approximated Bayesian cross-validation for a b3lmeta object
}
