% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.hmr.R
\name{diagnostic}
\alias{diagnostic}
\title{Generic diagnostic function.}
\usage{
diagnostic(object, ...)
}
\arguments{
\item{object}{The object generated by the function hmr.}

\item{...}{\dots}
}
\description{
Generic diagnostic function.
}
