% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.hmr.R
\name{effect}
\alias{effect}
\title{Generic effect function.}
\usage{
effect(object, ...)
}
\arguments{
\item{object}{The object generated by the function hmr.}

\item{...}{\dots}
}
\description{
Generic effect function.
}
