% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{fnrpcr}
\alias{fnrpcr}
\title{Meta-Analysis: Variation in False-Negative Rate of Reverse Transcriptase Polymerase Chain Reaction–Based SARS-CoV-2 Tests by Time Since Exposure}
\format{
A data frame with 410 rows and 11 columns. Each row represents the results from a study. The columns include:
\describe{
  \item{study}{Name of the author conducting the study.}
  \item{test}{Type of testing performed.}
  \item{day}{Number of days since symptom onset.}
  \item{day_min}{Minimum number of days since symptom onset. Applicable for studies by Guo et al. and Kim et al.}
  \item{day_max}{Maximum number of days since symptom onset. Applicable for studies by Guo et al. and Kim et al.}
  \item{n}{Total number of tests conducted on a given day.}
  \item{test_pos}{Number of positive test results.}
  \item{inconclusive}{Number of inconclusive test results. Applicable for studies by Kujawski et al. and Danis et al.}
  \item{nqp}{Number of positive but non-quantifiable test results, where the viral load is below the quantification threshold of log10(1) copies/1000 cells.}
  \item{pct_pos}{Proportion of positive tests expressed as a percentage.}
}
}
\source{
Kucirka LM, Lauer SA, Laeyendecker O, Boon D, Lessler J. Variation in False-Negative Rate of Reverse Transcriptase Polymerase Chain Reaction-Based SARS-CoV-2 Tests by Time Since Exposure. Ann Intern Med. 2020 Aug 18;173(4):262-267. doi: 10.7326/M20-1495. Epub 2020 May 13. PMID: 32422057; PMCID: PMC7240870.
}
\description{
A dataset summarizing the variation in false-negative rates of reverse transcriptase polymerase chain reaction (RT-PCR)–based SARS-CoV-2 tests
as a function of time since exposure.
}
\keyword{datasets}
