% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{healing}
\alias{healing}
\title{Efficacy of diabetic foot healing using adjuvant treatments}
\format{A matrix with 35 rows and 9 columns. Each row represents study results, 
the columns are:
   \describe{
    \item{Study}{Name of the first author and year.}
    \item{n_t}{Number of patients in the treatment group.}
    \item{n_c}{Number of patients in the control group.}
    \item{y_t}{Number of heal patients in the treatment group.}
    \item{y_c}{Number of heal patients in the control group.}
    \item{ndrop}{Total number of drop out patients.}
    \item{fup_weeks}{Length of followup in weeks.}
    \item{PAD}{Inclusion of patients with peripheral arterial disease.}
    \item{wagner_4}{Inclusion of patients with Wagner score 3 and 4.}
    }}
\source{
The data were obtainded from: Centre for Clinical Practice at NICE (UK and others) 
(2011), Clinical guideline 119. Diabetic foot problems: Inpatient Management of Diabetic Foot Problems.
 Tech. rep., National Institute for Health and Clinical Excellence.
}
\description{
Meta-analysis of 35 randomized controlled trials investigating the effectiveness
in the application of adjuvant therapies for diabetic patients compared to medical
routine care, where the endpoint was healing without amputations within a period
less than or equal to one year.
}
\references{
Verde, P.E. (2018) The Hierarchical Meta-Regression Approach and Learning 
from Clinical Evidence. Technical Report.
}
\keyword{datasets}
