% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{macula_rwe}
\alias{macula_rwe}
\title{Meta-analysis: 83 observational studies assessing the effectiveness of intravitreal therapy for diabetic maculaedema}
\format{
A dataframe with 83 rows and 13 columns. Each row represents study results, the columns are:
   \describe{
    \item{therapy}{Used Medication}
    \item{author_year}{Author and year.}
    \item{eyes}{Number of tested eyes.}
    \item{TE}{Mean Change in Visual Acuity after 12-Months. Where Visual Acuity is measured in VAR (Visual Acuity Rating) score}
    \item{seTE}{Standard Error of the Treatment Effect}
    \item{lower_95pct_ci}{Lower 95prc CI for TE}
    \item{upper_95pct_ci}{Upper 95prc CI for TE}
    \item{number_of_patients_at_baseline}{Number of Patients in Study at Baseline}
    \item{mean_age}{The mean age of patients per study}
    \item{baseline_va}{Mean Visual Acuity at Baseline. Where Visual Acuity is measured in VAR (Visual Acuity Rating) score}
    \item{12_month_va}{Mean Visual Acuity after 12 months. Where Visual Acuity is measured in VAR (Visual Acuity Rating) score}
    \item{baseline_cst}{Mean Central Subfield Thickness at Baseline}
    \item{12_month_cst}{Mean Central Subfield Thickness after 12 months}
    }
}
\source{
Mehta H, Nguyen V, Barthelmes D, Pershing S, Chi GC, Dopart P, Gillies MC. Outcomes of Over 40,000 Eyes Treated for Diabetic Macula Edema in Routine Clinical Practice: A Systematic Review and Meta-analysis. Adv Ther. 2022 Dec;39(12):5376-5390. doi: 10.1007/s12325-022-02326-8. Epub 2022 Oct 15. PMID: 36241963; PMCID: PMC9618488.
}
\description{
Meta-analysis of 83 studies comparing 12-month visual acuity change results in routine clinical practices
of intravitreal therapy for diabetic maculaedema (DME) to the change in RCTs by pooling data published in the
last decade on treated eyes of treatment effect from OS.
}
\keyword{datasets}
