% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b3lmeta.R, R/plot.b3lmeta.R
\name{plot.b3lmeta}
\alias{plot.b3lmeta}
\title{Generic plot function for bcmeta object in jarbes.}
\usage{
\method{plot}{b3lmeta}(
  x,
  x.lim = c(-3, 3),
  y.lim = c(0, 2.7),
  x.lab = "Treatment Effect: log(OR)",
  y.lab = "Posterior",
  title.plot.1 = "Mean Design Components",
  title.plot.2 = "Three Levels Bayesian Meta-Analysis",
  ...
)

\method{plot}{b3lmeta}(
  x,
  x.lim = c(-3, 3),
  y.lim = c(0, 2.7),
  x.lab = "Treatment Effect: log(OR)",
  y.lab = "Posterior",
  title.plot.1 = "Mean Design Components",
  title.plot.2 = "Three Levels Bayesian Meta-Analysis",
  ...
)
}
\arguments{
\item{x}{The object generated by the b3lmeta function.}

\item{x.lim}{Numeric vector of length 2 specifying the x-axis limits.}

\item{y.lim}{Numeric vector of length 2 specifying the y-axis limits.}

\item{x.lab}{Text with the label of the x-axis.}

\item{y.lab}{Text with the label of the y-axis.}

\item{title.plot.1}{Text for the posterior means by design.}

\item{title.plot.2}{Text for the posterior pooled mean.}

\item{...}{\dots}
}
\description{
Generic plot function for bcmeta object in jarbes.

Generic plot function for b3lmeta object in jarbes.
}
