% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dpmeta.R
\name{plot.dpmmeta}
\alias{plot.dpmmeta}
\title{Generic plot function for dpmeta object in jarbes.}
\usage{
\method{plot}{dpmmeta}(
  x,
  x.lim = c(-3, 3),
  y.lim = c(0, 2),
  x.lab = "Treatment Effect: log(OR)",
  y.lab = "Posterior",
  title.plot = "Simple Bayesian Meta-Analysis",
  ...
)
}
\arguments{
\item{x}{The object generated by the bmeta function.}

\item{x.lim}{Numeric vector of length 2 specifying the x-axis limits.}

\item{y.lim}{Numeric vector of length 2 specifying the y-axis limits.}

\item{x.lab}{Text with the label of the x-axis.}

\item{y.lab}{Text with the label of the y-axis.}

\item{title.plot}{Text for setting a title in the plot.}

\item{...}{\dots}
}
\description{
Generic plot function for dpmeta object in jarbes.
}
