% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ges.R
\name{plot.ges}
\alias{plot.ges}
\title{Generic plot function for ges object in jarbes.}
\usage{
\method{plot}{ges}(x, ...)
}
\arguments{
\item{x}{The object generated by the ges function.}

\item{...}{\dots}
}
\description{
Generic plot function for ges object in jarbes.
}
