% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gesmix.R
\name{plot.gesmix}
\alias{plot.gesmix}
\title{Generic plot function for metamixt object in jarbes.}
\usage{
\method{plot}{gesmix}(x, ...)
}
\arguments{
\item{x}{The object generated by the metamixt function.}

\item{...}{\dots}
}
\description{
Generic plot function for metamixt object in jarbes.
}
