% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmr.R
\name{plot.hmr}
\alias{plot.hmr}
\title{Generic plot function for hmr object in jarbes.}
\usage{
\method{plot}{hmr}(x, ...)
}
\arguments{
\item{x}{The object generated by the hmr function.}

\item{...}{\dots}
}
\description{
Generic plot function for hmr object in jarbes.
}
