% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metarisk.R
\name{plot.metarisk}
\alias{plot.metarisk}
\title{Generic plot function for metarisk object in jarbes.}
\usage{
\method{plot}{metarisk}(x, ...)
}
\arguments{
\item{x}{The object generated by the metarisk function.}

\item{...}{\dots}
}
\description{
Generic plot function for metarisk object in jarbes.
}
