% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.simData.R
\name{plot.simData}
\alias{plot.simData}
\title{Forest Plot for simData Objects}
\arguments{
\item{x}{A `simData` object created by `simData()`.}

\item{x.lim}{Numeric vector of length 2 specifying the x-axis limits.}

\item{x.lab}{Text with the label of the x-axis.}

\item{y.lab}{Text with the label of the y-axis.}

\item{title.plot}{Text for the title of the plot.}

\item{bias_colors}{Named character vector specifying colors for bias categories.}

\item{bias_legend_title}{Text label for the bias legend.}

\item{ref_line_color}{Color for the vertical reference line corresponding to the mean.}

\item{...}{Additional arguments (currently not used).}
}
\description{
Generates a forest plot of the simulated meta-analysis dataset,
showing the observed effect sizes (TE) with 95% confidence intervals.
}
