% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{ppvipd}
\alias{ppvipd}
\title{Efficacy of Pneumococcal Polysaccharide Vaccine in Preventing Invasive Pneumococcal Disease}
\format{A matrix with 11 rows and 6 columns. Each row represents study results, the columns are:
   \describe{
    \item{name}{Name of the first author and year.}
    \item{TE}{Treatment Effect as Log Odds Ratio.}
    \item{seTE}{Standard Error of the TE.}
    \item{n.v}{Number of patients in the vaccination group.}
    \item{n.c}{Number of patients in the control group.}
    \item{design}{Description of the study design.}
    }}
\source{
The data were obtainded from: Falkenhorst et al. (2017).
}
\description{
PPV23 (23-valent pneumococcal polysaccharide vaccine) with 3 Randomized Clinical Trials;
5 Cohort Studies and 3 Case-Control Studies.

The outcome variable IPD (Invasive Pneumococcal Disease).
}
\references{
Falkenhorst, G., Remschmidt, C., Harder, T., Hummers-Pradier, E., Wichmann, O., and Bogdan, C. (2017) Effectiveness of the 23-Valent Pneumococcal Polysaccharide Vaccine(PPV23) against Pneumococcal Disease in the Elderly: Systematic Review and Meta-Analysis. PLoS ONE 12(1): e0169368. doi:10.1371/journal.pone.0169368.

Verde, P.E. and Curcio, D. (2017) Hierarchical Meta-Regression Modelling: The Case of The Pneumococcal Polysaccharide Vaccine. Technical Report.
}
\keyword{datasets}
