% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b3lmeta.R
\name{print.b3lmeta}
\alias{print.b3lmeta}
\title{Generic print function for b3lmeta object in jarbes.}
\usage{
\method{print}{b3lmeta}(x, digits, ...)
}
\arguments{
\item{x}{The object generated by the function b3lmeta.}

\item{digits}{The number of significant digits printed. The default value is 3.}

\item{...}{\dots}
}
\description{
Generic print function for b3lmeta object in jarbes.
}
