% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcdpmeta.R
\name{print.bcdpmeta}
\alias{print.bcdpmeta}
\title{Generic print function for bcdpmeta object in jarbes.}
\usage{
\method{print}{bcdpmeta}(x, digits, ...)
}
\arguments{
\item{x}{The object generated by the function bcdpmeta.}

\item{digits}{The number of significant digits printed. The default value is 3.}

\item{...}{\dots}
}
\description{
Generic print function for bcdpmeta object in jarbes.
}
