% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcmixmeta.R
\name{print.bcmixmeta}
\alias{print.bcmixmeta}
\title{Generic print function for bcmixmeta object in jarbes.}
\usage{
\method{print}{bcmixmeta}(x, digits, ...)
}
\arguments{
\item{x}{The object generated by the function bcmixmeta.}

\item{digits}{The number of significant digits printed. The default value is 3.}

\item{...}{\dots}
}
\description{
Generic print function for bcmixmeta object in jarbes.
}
