% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmeta.R
\name{print.bmeta}
\alias{print.bmeta}
\title{Generic print function for bcmeta object in jarbes.}
\usage{
\method{print}{bmeta}(x, digits, ...)
}
\arguments{
\item{x}{The object generated by the function bcmeta.}

\item{digits}{The number of significant digits printed. The default value is 3.}

\item{...}{\dots}
}
\description{
Generic print function for bcmeta object in jarbes.
}
