% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpmeta.R
\name{print.dpmeta}
\alias{print.dpmeta}
\title{Generic print function for dpmeta object in jarbes.}
\usage{
\method{print}{dpmeta}(x, digits, ...)
}
\arguments{
\item{x}{The object generated by the function dpmeta.}

\item{digits}{The number of significant digits printed. The default value is 3.}

\item{...}{\dots}
}
\description{
Generic print function for dpmeta object in jarbes.
}
