% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpmetareg.R
\name{print.dpmetareg}
\alias{print.dpmetareg}
\title{Generic print function for dpmeta object in jarbes.}
\usage{
\method{print}{dpmetareg}(x, digits, ...)
}
\arguments{
\item{x}{The object generated by the function dpmmeta.}

\item{digits}{The number of significant digits printed. The default value is 3.}

\item{...}{\dots}
}
\description{
Generic print function for dpmeta object in jarbes.
}
