% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpmmeta.R
\name{print.dpmmeta}
\alias{print.dpmmeta}
\title{Generic print function for dpmmeta object in jarbes.}
\usage{
\method{print}{dpmmeta}(x, digits, ...)
}
\arguments{
\item{x}{The object generated by the function dpmmeta.}

\item{digits}{The number of significant digits printed. The default value is 3.}

\item{...}{\dots}
}
\description{
Generic print function for dpmmeta object in jarbes.
}
