% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ges.R
\name{print.ges}
\alias{print.ges}
\title{Generic print function for ges object in jarbes.}
\usage{
\method{print}{ges}(x, digits, ...)
}
\arguments{
\item{x}{The object generated by the function ges.}

\item{digits}{The number of significant digits printed. The default value is 3.}

\item{...}{\dots}
}
\description{
Generic print function for ges object in jarbes.
}
