% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gesmix.R
\name{print.gesmix}
\alias{print.gesmix}
\title{Generic print function for metamixt object in jarbes.}
\usage{
\method{print}{gesmix}(x, digits, ...)
}
\arguments{
\item{x}{The object generated by the function metamixt.}

\item{digits}{The number of significant digits printed. The default value is 3.}

\item{...}{\dots}
}
\description{
Generic print function for metamixt object in jarbes.
}
