% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmr.R
\name{print.hmr}
\alias{print.hmr}
\title{Generic print function for hmr object in jarbes.}
\usage{
\method{print}{hmr}(x, digits = 3, intervals = c(0.025, 0.25, 0.5,
  0.75, 0.975), ...)
}
\arguments{
\item{x}{The object generated by the function hmr.}

\item{digits}{The number of significant digits printed. The default value is 3.}

\item{intervals}{A numeric vector of probabilities with values in [0,1]. The default value is
intervals = c(0.025, 0.5, 0.975).}

\item{...}{\dots}
}
\description{
Generic print function for hmr object in jarbes.
}
