% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metarisk.R
\name{print.metarisk}
\alias{print.metarisk}
\title{Generic print function for metarisk object in jarbes.}
\usage{
\method{print}{metarisk}(x, digits, ...)
}
\arguments{
\item{x}{The object generated by the function metarisk.}

\item{digits}{The number of significant digits printed. The default value is 3.}

\item{...}{\dots}
}
\description{
Generic print function for metarisk object in jarbes.
}
