% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simData.R
\name{print.simData}
\alias{print.simData}
\title{Generic print function for simData object.}
\usage{
\method{print}{simData}(x, ...)
}
\arguments{
\item{x}{An object of class \code{simData} generated by the \code{simData} function.}

\item{...}{Additional arguments passed to \code{print.data.frame}.}
}
\description{
Generic print function for simData object.
}
