% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.b3lmeta.R
\name{summary.b3lmeta}
\alias{summary.b3lmeta}
\title{Generic summary function for bmeta object in jarbes}
\usage{
\method{summary}{b3lmeta}(object, digits = 3, ...)
}
\arguments{
\item{object}{The object generated by the bmeta function.}

\item{digits}{The number of significant digits printed. The default value is 3.}

\item{...}{\dots}
}
\description{
Generic summary function for bmeta object in jarbes
}
