% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bcbnp.R
\name{summary.bcbnp}
\alias{summary.bcbnp}
\title{Generic summary function for bcbnp object in jarbes.}
\usage{
\method{summary}{bcbnp}(object, digits = 3, show.clustering = FALSE, show.neff = FALSE, ...)
}
\arguments{
\item{object}{The object generated by the bcbnp function.}

\item{digits}{The number of significant digits printed. The default value is 3.}

\item{show.clustering}{A logical value indicating whether to show the clustering parameters table. Defaults to `TRUE`.}

\item{show.neff}{A logical value indicating whether to show the effective sample size column (n.eff) and Rhat. Defaults to `TRUE`.}

\item{...}{\dots}
}
\description{
Generic summary function for bcbnp object in jarbes.
}
