% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bcdpmeta.R
\name{summary.bcdpmeta}
\alias{summary.bcdpmeta}
\title{Generic summary function for bcdpmeta object in jarbes}
\usage{
\method{summary}{bcdpmeta}(object, digits = 3, ...)
}
\arguments{
\item{object}{The object generated by the bcmeta function.}

\item{digits}{The number of significant digits printed. The default value is 3.}

\item{...}{\dots}
}
\description{
Generic summary function for bcdpmeta object in jarbes
}
