% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bcmeta.R
\name{summary.bcmeta}
\alias{summary.bcmeta}
\title{Generic summary function for bcmeta object in jarbes}
\usage{
\method{summary}{bcmeta}(object, digits = 3, ...)
}
\arguments{
\item{object}{The object generated by the bcmeta function.}

\item{digits}{The number of significant digits printed. The default value is 3.}

\item{...}{\dots}
}
\description{
Generic summary function for bcmeta object in jarbes
}
