% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.dpmeta.R
\name{summary.dpmeta}
\alias{summary.dpmeta}
\title{Generic summary function for dpmmeta object in jarbes}
\usage{
\method{summary}{dpmeta}(object, digits = 3, ...)
}
\arguments{
\item{object}{The object generated by the dmpmeta function.}

\item{digits}{The number of significant digits printed. The default value is 3.}

\item{...}{\dots}
}
\description{
Generic summary function for dpmmeta object in jarbes
}
