% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.metarisk.R
\name{summary.metarisk}
\alias{summary.metarisk}
\title{Generic summary function for metarisk object in jarbes}
\usage{
\method{summary}{metarisk}(object, digits = 3, ...)
}
\arguments{
\item{object}{The object generated by the metarisk function.}

\item{digits}{The number of significant digits printed. The default value is 3.}

\item{...}{\dots}
}
\description{
Generic summary function for metarisk object in jarbes
}
