% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.simData.R
\name{summary.simData}
\alias{summary.simData}
\title{Summary method for simData objects}
\usage{
\method{summary}{simData}(object, digits = 3, ...)
}
\arguments{
\item{object}{A simData object created by simData().}

\item{digits}{The number of significant digits printed. The default value is 3.}

\item{...}{Additional arguments (currently not used).}
}
\value{
A formatted summary of the meta-analysis simulation.
}
\description{
Provides a summary of the simulated meta-analysis dataset using the known simulation parameters.
}
