% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{trisomy21}
\alias{trisomy21}
\title{Meta-analysis: Observational studies assessing the relationship of
a positive ICPC (Isolated Choroid Plexus Cyst) on Trisomy 21}
\format{
A dataframe with 22 rows and 6 columns. Each row represents study results, the columns are:
   \describe{
    \item{year}{Year of publication.}
    \item{author}{Principal author of the publication.}
    \item{y}{Number of cases of ICPC with Trisomy 21.}
    \item{n}{Total number o cases with ICPC.}
    \item{mean.GA}{Mean gestational time in weeks.}
    \item{study.design}{Study design: prospective or retrospective cohort.}
    }
}
\source{
Kürten C, Knippel A, Verde P, Kozlowski P. A Bayesian risk analysis for Trisomy 21 in isolated choroid plexus cyst: combining a prenatal database with a meta-analysis. J Matern Fetal Neonatal Med. 2019 Jun 11:1-9. doi: 10.1080/14767058.2019.1622666. Epub ahead of print. PMID: 31113245.
}
\description{
Meta-analysis of 22 Observational Studies from PubMed, Cochrane Library and SciELO databases
that assessed the relationship of a positive ICPC (Isolated Choroid Plexus Cyst) on Trisomy 21
}
\keyword{datasets}
