% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bmeta.R
\name{summary.bmeta}
\alias{summary.bmeta}
\title{Generic summary function for bmeta object in jarbes}
\usage{
\method{summary}{bmeta}(object, digits = 3, ...)
}
\arguments{
\item{object}{The object generated by the bmeta function.}

\item{digits}{The number of significant digits printed. The default value is 3.}

\item{...}{\dots}
}
\description{
Generic summary function for bmeta object in jarbes
}
