% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.hmr.R
\name{summary.hmr}
\alias{summary.hmr}
\title{Generic summary function for hmr object in jarbes}
\usage{
\method{summary}{hmr}(object, digits = 3, ...)
}
\arguments{
\item{object}{The object generated by the hmr function.}

\item{digits}{The number of significant digits printed. The default value is 3.}

\item{...}{\dots}
}
\description{
Generic summary function for hmr object in jarbes
}
